# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Unifesto Authentication & User API
  description: Mobile-number-centric identity system for Unifesto
  version: 3.0.0
  contact:
    name: Unifesto Team

servers:
  - url: https://api.unifesto.app
    description: Production server
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Users
    description: User profile management
  - name: Roles
    description: Role management

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mobileNumber:
          type: string
          example: "+919876543210"
        mobileVerified:
          type: boolean
        username:
          type: string
          nullable: true
          example: "abhinavtej"
        fullName:
          type: string
          nullable: true
          example: "Abhinav Tej"
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        instagramUrl:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        isOnboarded:
          type: boolean
        createdAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
        requiresMobileVerification:
          type: boolean
        tempToken:
          type: string
          description: Present only when mobile verification is required

    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          enum: [ADMIN, SUPER_ORGANISER, ORGANISER, CO_ORGANISER, MEMBER]
        name:
          type: string
        scope:
          type: string
          enum: [platform, space]
        createdAt:
          type: string
          format: date-time

    UserRole:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
          nullable: true
        assignedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time

    UserRoleDetails:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
        roleCode:
          type: string
        roleName:
          type: string
        roleScope:
          type: string
        spaceId:
          type: string
          format: uuid
          nullable: true
        assignedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string

paths:
  /auth/google:
    post:
      tags:
        - Auth
      summary: Login with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idToken
              properties:
                idToken:
                  type: string
                  description: Google ID token
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/apple:
    post:
      tags:
        - Auth
      summary: Login with Apple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identityToken
                - authorizationCode
              properties:
                identityToken:
                  type: string
                authorizationCode:
                  type: string
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed

  /auth/email:
    post:
      tags:
        - Auth
      summary: Send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Failed to send OTP

  /auth/email/verify:
    post:
      tags:
        - Auth
      summary: Verify email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid OTP

  /auth/mobile/send-otp:
    post:
      tags:
        - Auth
      summary: Send OTP to mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mobileNumber
                - tempToken
              properties:
                mobileNumber:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                  example: "+919876543210"
                tempToken:
                  type: string
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Failed to send OTP

  /auth/verify-mobile:
    post:
      tags:
        - Auth
      summary: Verify mobile number with OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mobileNumber
                - otp
                - tempToken
              properties:
                mobileNumber:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                otp:
                  type: string
                tempToken:
                  type: string
      responses:
        '200':
          description: Mobile verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid OTP

  /auth/session:
    get:
      tags:
        - Auth
      summary: Get current session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized

  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

    patch:
      tags:
        - Users
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  pattern: '^[a-z0-9_]{3,50}$'
                fullName:
                  type: string
                  maxLength: 255
                bio:
                  type: string
                  maxLength: 1000
                linkedinUrl:
                  type: string
                  format: uri
                instagramUrl:
                  type: string
                  format: uri
                githubUrl:
                  type: string
                  format: uri
                websiteUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '409':
          description: Username already taken

  /users/me/onboard:
    post:
      tags:
        - Users
      summary: Complete onboarding
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /users/me/avatar:
    post:
      tags:
        - Users
      summary: Upload avatar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - avatar
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Image file (max 5MB, jpg/jpeg/png/webp)
      responses:
        '200':
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatarUrl:
                    type: string
        '401':
          description: Unauthorized
        '400':
          description: Invalid file

  /users/check-username:
    post:
      tags:
        - Users
      summary: Check username availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  pattern: '^[a-z0-9_]{3,50}$'
      responses:
        '200':
          description: Username availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean

  /users/{username}:
    get:
      tags:
        - Users
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

  /roles:
    get:
      tags:
        - Roles
      summary: Get all available roles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized

  /roles/users/{userId}:
    get:
      tags:
        - Roles
      summary: Get user roles
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserRoleDetails'
        '401':
          description: Unauthorized

  /roles/assign:
    post:
      tags:
        - Roles
      summary: Assign role to user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - roleId
              properties:
                userId:
                  type: string
                  format: uuid
                roleId:
                  type: string
                  format: uuid
                spaceId:
                  type: string
                  format: uuid
                  description: Required for space roles, must be null for platform roles
      responses:
        '200':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRole'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Role not found
        '409':
          description: Role already assigned

  /roles/{userRoleId}:
    delete:
      tags:
        - Roles
      summary: Remove role from user
      security:
        - BearerAuth: []
      parameters:
        - name: userRoleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized

  /roles/check/{userId}/{roleCode}:
    get:
      tags:
        - Roles
      summary: Check if user has specific role
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: roleCode
          in: path
          required: true
          schema:
            type: string
            enum: [ADMIN, SUPER_ORGANISER, ORGANISER, CO_ORGANISER, MEMBER]
      responses:
        '200':
          description: Role check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasRole:
                    type: boolean
        '401':
          description: Unauthorized